Introduce realloc(...)

+++ src/usr/include/string.h
@@ -12,7 +12,7 @@ int strcmpi(const char *l, const char *r);
 int strlen(const char *str);
 char *strcpy(char *dest, const char *src);
 char *strncpy(char *dest, const char *src, size_t n);
-void *memcpy(void *dest, const void *src, unsigned int n);
+void *memcpy(void *dest, const void *src, size_t n);
 void *memset(void *dest, const unsigned char c, size_t n);
 
 char *strchr(const char *str, char ch);
